Hello, world!
Add runtime and end-to-end test

Add a runtime library to link with generated modules.
 * It goes in the host/ directory to match the wasm spec interpreter.
 * Add a host-provided stdio module with a simple integer print function
 * Add a test main function that calls an export from a wasm module.

Add a test compiler driver that calls waot to compile a wast file to llvm,
translates the result with llc, and links that with the runtime to
create an executable.

Add a test that pulls it all together.