Properly support multi-module for LLVM output
LLVM assumes one module per output file, so we can't just do what we did
with the AST files. Instead we can support multiple wasm modules per
LLVM module. This works because public names (imports and exports) are
already mangled with the module name and internal names are
automatically fixed up by LLVM when they clash.
Also change the waot llvm visitor to take the LLVM module in the
constructor (and not own it) rather than create it on module
construction and return it from the module visitor.