Store modules as pointers in the parser
wasm::Modules were stored in a vector in Parser, but we were taking a
pointer to them, and they were subsequently getting reallocated when the
vector was enlarged, causing the pointers to be stale.
Instead use a vector of unique_ptrs, so that we still get automatic
deallocation but we can also take raw pointers.

Also use it for Segments in the module. Future changes will do it for
the other types too.