Add __assert_fail runtime functions
Instead of just trapping on assert_eq failure, call a runtime function
that prints an error message and returns. It also sets a global variable
that is returned by the script main function so that the whole program
exits with an error status.

Also fix assert_eq with f64 arguments and add more tests cases.